squash! test/FabricTransport/NativeMessageStreamTest.cs
Implement a fix for a build error in `test/FabricTransport/NativeMessageStreamTest.cs`.

The `ReadByte.ReturnsBytesSequentiallyAcrossNativeBuffers` test contains `Assert.Equal((int)expectedBytes[i], sut.ReadByte());`. The `(int)` cast on the `byte` element triggers analyzer error `IDE0004: Cast is redundant`, which fails the build. Remove the redundant cast while preserving the intent of comparing the `int` result of `ReadByte()` directly (so the `-1` end-of-stream sentinel cannot be masked as `255`). Build the test project and run the tests to verify they pass.

+++ test/FabricTransport/NativeMessageStreamTest.cs
@@ -249,7 +249,7 @@ public abstract class NativeMessageStreamTest: IDisposable
         public void ReturnsBytesSequentiallyAcrossNativeBuffers()
         {
             for (int i = 0; i < expectedBytes.Length; i++)
-                Assert.Equal((int)expectedBytes[i], sut.ReadByte());
+                Assert.Equal(expectedBytes[i], sut.ReadByte());
         }
 
         [Fact]
